refactor(lodex): Gather runQuery and reduceQuery tests
Because they both use Mongo, and tearing up mongo has a cost.
